+++ CHANGELOG.md
@@ -39,7 +39,8 @@ All user visible changes to this project will be documented in this file. This p
     - Send relay mode in `Event::PeerCreated` which is used for configuring client's `RtcIceTransportPolicy` ([#79]);
     - Emit `TracksApplied` event to create new and update existing tracks ([#105]);
     - `PeerConnection` renegotiation functionality ([#105]);
-    - Calculate and send call quality score based on RTC stats ([#132]).
+    - Calculate and send call quality score based on RTC stats ([#132]);
+    - Muting/unmuting `MediaTrack`s by receiver ([#127]).
 - [Coturn] integration:
     - [Coturn] sessions destroying ([#84]);
     - [Coturn] stats processing ([#94]).
@@ -75,6 +76,7 @@ All user visible changes to this project will be documented in this file. This p
 [#95]: /../../pull/95
 [#98]: /../../pull/98
 [#105]: /../../pull/105
+[#127]: /../../pull/127
 [#132]: /../../pull/132
 [#135]: /../../pull/135