+++ jason/CHANGELOG.md
@@ -48,7 +48,12 @@ All user visible changes to this project will be documented in this file. This p
         - `Room.on_close` callback for WebSocket close initiated by server ([#55]);
         - `RemoteMediaStream.on_track_enabled` and `RemoteMediaStream.on_track_disabled` callbacks being called when `MediaTrack` is enabled or disabled ([#123]);
         - `RemoteMediaStream.on_track_added` callback being called when new receiver `MediaTrack` is added ([#123]);
-        - `RemoteMediaStream.has_active_audio` and `RemoteMediaStream.has_active_video` methods returning current state of the receivers ([#123]).
+        - `RemoteMediaStream.has_active_audio` and `RemoteMediaStream.has_active_video` methods returning current state of the receivers ([#123]);
+        - Muting/unmuting remote video/audio ([#127]):
+            - `Room.mute_remote_audio`;
+            - `Room.unmute_remote_audio`;
+            - `Room.mute_remote_video`;
+            - `Room.unmute_remote_video`.
     - Optional tracks support ([#106]);
     - `RtcIceTransportPolicy` configuration ([#79]).
 - Room management:
@@ -96,6 +101,7 @@ All user visible changes to this project will be documented in this file. This p
 [#120]: /../../pull/120
 [#123]: /../../pull/123
 [#124]: /../../pull/124
+[#127]: /../../pull/127
 [#132]: /../../pull/132
 [#137]: /../../pull/137